整理代码
Former-commit-id: 407deffdd1736a5d018c81995cf118a750a8f510 [formerly f526352ec357b8a0f8ba26ed216af7e36c415578]
Former-commit-id: 6cde6fbe89a18c7d7117d6542eea79fe5ae9d357

+++ faber-admin/src/main/java/com/faber/msg/rest/MsgController.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/admin/msg")
-public class MsgController extends BaseController<MsgBiz, Msg, Integer> {
+public class MsgController extends BaseController<MsgBiz, Msg, Long> {
 
     /**
      * 消息数量统计。
